feat: drop files anywhere, any tab
