--- components/InfoPane/FileUploadControl.scss
@@ -1,15 +1,0 @@
-#drop_target {
-  border: dashed 2px var(--text-dim);
-  border-radius: var(--rad_sm);
-  background-color: var(--bg2);
-  padding: 4em 1em;
-}
-
-#drop_target.drag {
-  border-color: #68a;
-  background-color: #ccf;
-}
-
-#drop_target * {
-  pointer-events: none;
-}
