+++ index.scss
@@ -141,3 +141,8 @@ kbd {
   bottom: 0;
   left: 0;
 }
+
+.link {
+  color: var(--accent);
+  text-decoration: underline;
+}
